# Tools/BackendGenerator/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Events API",
    "version": "v1"
  },
  "paths": {
    "/api/Auth/Register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "api/Auth/Register_POST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDetails"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDetails"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserDetails"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/Login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "api/Auth/Login_POST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCredentials"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCredentials"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCredentials"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/Logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "api/Auth/Logout_POST",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/GetAllUsers": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "api/Auth/GetAllUsers_GET",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/Delete/{name}": {
      "delete": {
        "tags": [
          "Auth"
        ],
        "operationId": "api/Auth/Delete/{name}_DELETE",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Event": {
      "get": {
        "tags": [
          "Event"
        ],
        "operationId": "api/Event_GET",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Event"
        ],
        "operationId": "api/Event_POST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Event/{id}": {
      "get": {
        "tags": [
          "Event"
        ],
        "operationId": "api/Event/{id}_GET",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Event"
        ],
        "operationId": "api/Event/{id}_DELETE",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Event/{Id}": {
      "put": {
        "tags": [
          "Event"
        ],
        "operationId": "api/Event/{Id}_PUT",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/EnableEvent": {
      "post": {
        "tags": [
          "Event"
        ],
        "operationId": "EnableEvent_POST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnableEventInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EnableEventInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EnableEventInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/DisabledEvents": {
      "get": {
        "tags": [
          "Event"
        ],
        "operationId": "DisabledEvents_GET",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnableEventInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EnableEventInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EnableEventInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Teste/HelloWord": {
      "get": {
        "tags": [
          "Teste"
        ],
        "operationId": "api/Teste/HelloWord_GET",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Teste/testStrings": {
      "get": {
        "tags": [
          "Teste"
        ],
        "operationId": "api/Teste/testStrings_GET",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Teste/GetTesteStringByIndex": {
      "post": {
        "tags": [
          "Teste"
        ],
        "operationId": "api/Teste/GetTesteStringByIndex_POST",
        "parameters": [
          {
            "name": "i",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Teste/UpdateTesteStringByIndex": {
      "post": {
        "tags": [
          "Teste"
        ],
        "operationId": "api/Teste/UpdateTesteStringByIndex_POST",
        "parameters": [
          {
            "name": "i",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "newStr",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Teste": {
      "delete": {
        "tags": [
          "Teste"
        ],
        "operationId": "api/Teste_DELETE",
        "parameters": [
          {
            "name": "i",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Ticket/Buy": {
      "post": {
        "tags": [
          "Ticket"
        ],
        "operationId": "api/Ticket/Buy_POST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyTicketInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyTicketInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BuyTicketInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BuyTicketInput": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "integer",
            "format": "int32"
          },
          "userEmail": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EnableEventInput": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Event": {
        "required": [
          "capacity",
          "createdBy",
          "createdOn",
          "endDate",
          "location",
          "name",
          "photo",
          "startDate",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "minLength": 1,
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "minLength": 1,
            "type": "string"
          },
          "updatedBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          },
          "updatedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "photo": {
            "minLength": 1,
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "location": {
            "minLength": 1,
            "type": "string"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "tickets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ticket"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "LoginCredentials": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Ticket": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idEvent": {
            "type": "integer",
            "format": "int32"
          },
          "event": {
            "$ref": "#/components/schemas/Event"
          },
          "ticketBuyer": {
            "$ref": "#/components/schemas/TicketBuyers"
          }
        },
        "additionalProperties": false
      },
      "TicketBuyers": {
        "required": [
          "userEmail"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userEmail": {
            "minLength": 1,
            "type": "string"
          },
          "vatNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "boughtOn": {
            "type": "string",
            "format": "date-time"
          },
          "ticketId": {
            "type": "integer",
            "format": "int32"
          },
          "ticket": {
            "$ref": "#/components/schemas/Ticket"
          }
        },
        "additionalProperties": false
      },
      "UserDetails": {
        "required": [
          "email",
          "password",
          "userName"
        ],
        "type": "object",
        "properties": {
          "userName": {
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    }
  }
}